helpers: passhash moved
